block/throttle-groups: fix throttle properties typename
The getter and setter use visit_type_int64, not visit_type_int.

Fixes: 432d889e55e2 ("block: convert ThrottleGroup to object with QOM")
Message-Id: <20260818-qom-qapi-v3-34-24b8bbbe3d86@redhat.com>

+++ block/throttle-groups.c
@@ -985,7 +985,7 @@ static void throttle_group_obj_class_init(ObjectClass *klass,
     for (i = 0; i < sizeof(properties) / sizeof(ThrottleParamInfo); i++) {
         object_class_property_add(klass,
                                   properties[i].name,
-                                  "int",
+                                  "int64",
                                   throttle_group_get,
                                   throttle_group_set,
                                   NULL, &properties[i]);
